requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
neo4j==5.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
setuptools==69.0.3
wheel==0.42.0